fix build error

+++ cgotorch/cgotorch.h
@@ -118,7 +118,7 @@ const char *MaxPool2d(Tensor input, int64_t *kernel_data, int64_t kernel_len,
                       int64_t *stride_data, int64_t stride_len,
                       int64_t *padding_data, int64_t padding_len,
                       int64_t *dilation_data, int64_t dilation_len,
-                      int64_t ceil_mode, Tensor *result);
+                      int8_t ceil_mode, Tensor *result);
 
 const char *AdaptiveAvgPool2d(Tensor input, int64_t *output_size_data,
                               int64_t output_size_len, Tensor *result);
